Move some of main-frontend course id components
