+++ services/main-frontend/src/components/page-specific/manage/courses/id/change-request/EditProposalView.tsx
@@ -8,8 +8,8 @@ import {
   BlockProposalAction,
   BlockProposalInfo,
   PageProposal,
-} from "../shared-module/bindings"
-import Button from "../shared-module/components/Button"
+} from "../../../../../../shared-module/bindings"
+import Button from "../../../../../../shared-module/components/Button"
 
 export interface Props {
   proposal: PageProposal
