+++ shared-module/src/locales/en/main-frontend.json
@@ -21,6 +21,7 @@
   "error-could-not-find-edit-history-for-page": "Could not find any edit history for the page",
   "error-loading-organizations": "Error loading organizations.",
   "error-title": "Error",
+  "error-unknown-in-component": "Unknown error when rendering {{component}}",
   "export-points-as-csv": "Export points as csv",
   "feedback-given": "Feedback: {{feedback-given}}",
   "selected-text": "Selected text: {{selected-text}}",
