+++ shared-module/src/locales/fi/main-frontend.json
@@ -21,6 +21,7 @@
   "error-could-not-find-edit-history-for-page": "Sivulle ei löytynyt muokkaushistoriaa",
   "error-loading-organizations": "Virhe organisaatioiden lataamisessa",
   "error-title": "Virhe",
+  "error-unknown-in-component": "Määrittelemätön virhe komponentissa {{component}}",
   "export-points-as-csv": "Lataa pisteet CSV-tiedostona",
   "feedback-given": "Palaute: {{feedback-given}}",
   "selected-text": "Valittu teksti: {{selected-text}}",
